requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
airflow-provider-great-expectations

openlineage-airflow

pyspark~=3.2.1
safety
pyan3
